Cleanup: factor stemming class from terminology
